# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "抽奖系统后端API",
    "description": "一个支持多种抽奖模式的完整抽奖系统API，包含认证、活动管理、参与者管理、抽奖功能、系统管理等模块。",
    "version": "1.0.0",
    "contact": {
      "name": "开发团队",
      "email": "buduan461@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "开发环境"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "webhookAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Webhook Token"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "用户ID"
          },
          "username": {
            "type": "string",
            "minLength": 3,
            "maxLength": 50,
            "description": "用户名"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "邮箱"
          },
          "role": {
            "type": "string",
            "enum": ["super_admin", "admin"],
            "description": "用户角色"
          },
          "status": {
            "type": "string",
            "enum": ["active", "inactive"],
            "description": "账户状态"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          }
        }
      },
...
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "请求是否成功"
          },
          "data": {
            "type": "object",
            "description": "响应数据"
          },
          "message": {
            "type": "string",
            "description": "响应消息"
          }
        },
        "required": ["success"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "错误代码"
              },
              "message": {
                "type": "string",
                "description": "错误描述"
              },
              "details": {
                "type": "string",
                "description": "详细错误信息"
              }
            }
          }
        }
      }
    },
    "responses": {
      "UnauthorizedError": {
        "description": "未认证",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "权限不足",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "ValidationError": {
        "description": "请求参数错误",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  },
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["认证模块"],
        "summary": "用户登录",
        "description": "用户登录获取访问令牌",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 50,
                    "description": "用户名"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 6,
                    "description": "密码"
                  }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "token": {
                              "type": "string",
                              "description": "JWT访问令牌"
                            },
                            "user": {
                              "$ref": "#/components/schemas/User"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": ["认证模块"],
        "summary": "创建新管理员账户",
        "description": "创建新管理员账户（仅超级管理员可操作）",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 50,
                    "pattern": "^[a-zA-Z0-9_]+$",
                    "description": "用户名"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "邮箱"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 6,
                    "pattern": "^(?=.*[A-Za-z])(?=.*\\d)",
                    "description": "密码（必须包含字母和数字）"
                  },
                  "role": {
                    "type": "string",
                    "enum": ["admin", "super_admin"],
                    "description": "用户角色"
                  }
                },
                "required": ["username", "email", "password", "role"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "用户创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "user": {
                              "$ref": "#/components/schemas/User"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["认证模块"],
        "summary": "获取当前用户信息",
        "description": "获取当前登录用户的信息",
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "user": {
                              "$ref": "#/components/schemas/User"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/auth/password": {
      "put": {
        "tags": ["认证模块"],
        "summary": "修改当前用户密码",
        "description": "修改当前用户密码",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "old_password": {
                    "type": "string",
                    "description": "原密码"
                  },
                  "new_password": {
                    "type": "string",
                    "minLength": 6,
                    "pattern": "^(?=.*[A-Za-z])(?=.*\\d)",
                    "description": "新密码（必须包含字母和数字）"
                  }
                },
                "required": ["old_password", "new_password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "密码修改成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
...
  },
...
}